requirements.txt
alabaster==0.7.12
Babel==2.8.0
boto==2.49.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
config==0.5.0.post0
cryptography==3.1
decorator==4.4.2
docutils==0.16
futures==3.1.1
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tornado==6.0.4
urllib3==1.25.10
xmltodict==0.12.0
yandex-speech==0.4
